ShopMicro-FASE1/product-service/requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
redis==5.0.1